requirements.txt
numpy~=1.22.4
tensorflow~=2.7.3
tensorflow-gpu~=2.7.3
six~=1.16.0
gdown~=4.4.0
tqdm~=4.62.3
setuptools~=62.5.0
sounddevice~=0.4.4
scipy~=1.8.1
PyYAML~=6.0
Soran~=0.0.0
nltk~=3.6.7
matplotlib~=3.4.3
pip~=22.1.2
markovify~=0.9.4
psycopg2~=2.9.3
future~=0.18.2
Pillow~=8.4.0
espeakng~=1.0.2
playsound~=1.3.0
wget~=3.2
pyserial~=3.5